// src/common/enum/shipment-status.enum.ts
export enum ShipmentStatus {
    READY_TO_PICKUP = 'READY_TO_PICKUP',
    WAITING_PICKUP = 'WAITING_PICKUP',
    PICKED_UP = 'PICKED_UP',
    IN_TRANSIT = 'IN_TRANSIT',
    ARRIVED_AT_BRANCH = 'ARRIVED_AT_BRANCH',
    AT_BRANCH = 'AT_BRANCH',
    DEPARTED_FROM_BRANCH = 'DEPARTED_FROM_BRANCH',
    READY_TO_PICKUP_AT_BRANCH = 'READY_TO_PICKUP_AT_BRANCH',
    DELIVERED = 'DELIVERED',
    COMPLETED = 'COMPLETED',
    ON_THE_WAY = 'ON_THE_WAY',
    ON_THE_WAY_TO_ADDRESS = 'ON_THE_WAY_TO_ADDRESS',
    READY_TO_DELIVER = 'READY_TO_DELIVER',
}